# Speed up serializing strings, re-use hashes of struct zend_string