--- a/src/php7/hash.h
+++ b/src/php7/hash.h
@@ -19,6 +19,7 @@
 #endif
 
 #include <stddef.h>
+#include "zend_types.h"
 
 /** Key/value pair of hash_si.
  * @author Oleg Grenrus <oleg.grenrus@dynamoid.com>
@@ -94,12 +95,11 @@ int hash_si_find (struct hash_si *h, const char *key, size_t key_len, uint32_t *
 /** Finds value from hash_si.
  * Value returned thru value param.
  * @param h Pointer to hash_si struct.
- * @param key Pointer to key.
- * @param key_len Key length.
+ * @param key zend_string with key
  * @param[out] value Found value.
  * @return 0 if found, 1 if not.
  */
-struct hash_si_result hash_si_find_or_insert(struct hash_si *h, const char *key, size_t key_len, uint32_t value);
+struct hash_si_result hash_si_find_or_insert(struct hash_si *h, zend_string *key, uint32_t value);
 
 /** Remove value from hash_si.
  * Removed value is available thru value param.