with CFLAGS=-O3

serialize-objectarray sped up from 0.239 to 0.215 seconds
serialize-stringarray sped up from 0.317 to 0.225 seconds